backend/requirements.txt
fastapi==0.115.*
uvicorn[standard]==0.34.*
httpx==0.28.*
google-generativeai==0.8.*
pydantic==2.*
pydantic-settings==2.*
python-dotenv==1.*
PyJWT[crypto]==2.*
